refactor(test): build-tag OCP scheme registrations and drop legacy SetupEnvTest (#1329)
* refactor: guard OCP scheme registrations with distro build tags

Prometheus Operator monitoring APIs, OpenShift Route, and Istio v1
scheme registrations were wired unconditionally, pulling OCP-only
dependencies into upstream builds. This makes the upstream binary carry
dead scheme registrations and pulls in imports that don't belong there.

Uses the established companion file pattern (//go:build distro /
!distro) to gate these registrations, keeping the upstream build surface
clean while preserving full functionality in distro builds.


* refactor(test): migrate credentials suite to NewEnvTest, drop SetupEnvTest

The credentials suite was the last caller of the legacy SetupEnvTest
path. Migrate it to NewEnvTest().BuildEnvironment(), which uses a fresh
per-suite scheme (t.Scheme) instead of mutating the global scheme.Scheme.

With no callers left, remove SetupEnvTest and its build-tag companion
files (envtest_setup_default.go / envtest_setup_ocp.go), collapsing the
two test setup paths into one.


* simplify: drop manual capacity hints in scheme slice construction

Replace make+append pairs with the append([]T{...}, rest...) idiom.
The magic number capacity hints (5+len(distroSchemes), 8+len(...), etc.)
silently drift whenever entries are added or removed. The composite
literal approach allocates exactly once with no number to maintain.


* fix: wire GOTAGS through Dockerfile and docker-build for main controller

routev1.AddToScheme was moved behind //go:build distro in
main_schemes_ocp.go but the main Dockerfile had no ARG GOTAGS and the
docker-build Makefile target was not passing --build-arg GOTAGS.
This meant the binary was always compiled without the distro tag,
main_schemes_default.go (no-op) was linked, and v1.Route was never
registered - causing the InferenceGraph controller to crash on startup.

Follow the pattern already in place for llmisvc-controller.Dockerfile
and docker-build-llmisvc.


* docs(rules): extend build-tags rule to cover Dockerfiles and Makefile targets

Adds rules 7 and 8 to catch the class of bug where a Dockerfile builds
a Go binary that uses //go:build distro code but does not declare
ARG GOTAGS or pass -tags to go build - and the corresponding Makefile
target does not pass --build-arg GOTAGS.

The canonical references are llmisvc-controller.Dockerfile (correct)
and the main Dockerfile before this fix (incorrect).


---------
