+++ cmd/manager/main_schemes_ocp.go
@@ -1,5 +1,7 @@
+//go:build distro
+
 /*
-Copyright 2023 The KServe Authors.
+Copyright 2026 The KServe Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,20 +16,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package testing
+package main
 
 import (
-	"path/filepath"
-
-	kservescheme "github.com/kserve/kserve/pkg/scheme"
+	routev1 "github.com/openshift/api/route/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 )
 
-// NewEnvTest prepares k8s EnvTest with prereq
-func NewEnvTest(options ...Option) *Config {
-	testCRDs := WithCRDs(
-		filepath.Join(ProjectRoot(), "test", "crds"),
-	)
-	schemes := WithScheme(append([]AddToSchemeFunc{kservescheme.AddAll}, additionalTestSchemes()...)...)
-
-	return Configure(append(options, testCRDs, schemes)...)
+// registerDistroSchemes registers OCP-specific API schemes (e.g. OpenShift Routes).
+func registerDistroSchemes(s *runtime.Scheme) error {
+	return routev1.AddToScheme(s)
 }
