+++ pkg/testing/ctrl_default.go
@@ -1,5 +1,7 @@
+//go:build !distro
+
 /*
-Copyright 2023 The KServe Authors.
+Copyright 2026 The KServe Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -16,18 +18,8 @@ limitations under the License.
 
 package testing
 
-import (
-	"path/filepath"
-
-	kservescheme "github.com/kserve/kserve/pkg/scheme"
-)
-
-// NewEnvTest prepares k8s EnvTest with prereq
-func NewEnvTest(options ...Option) *Config {
-	testCRDs := WithCRDs(
-		filepath.Join(ProjectRoot(), "test", "crds"),
-	)
-	schemes := WithScheme(append([]AddToSchemeFunc{kservescheme.AddAll}, additionalTestSchemes()...)...)
-
-	return Configure(append(options, testCRDs, schemes)...)
+// additionalTestSchemes returns no extra scheme registration functions
+// for upstream builds.
+func additionalTestSchemes() []AddToSchemeFunc {
+	return nil
 }
